+++ apps/mobile/src/features/library/shared/DataFetchingError.tsx
@@ -1,5 +1,7 @@
 import { StyleSheet, View } from 'react-native'
-import { Button, Text, useTheme } from 'react-native-paper'
+import { Text, useTheme } from 'react-native-paper'
+
+import Button from '@/components/common/Button'
 
 interface DataFetchingErrorProps {
 	text?: string